requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
celery
redis
faiss-cpu
sentence-transformers
python-docx
pymupdf
openai
pydantic
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
celery
redis
faiss-cpu
sentence-transformers
python-docx
pymupdf
openai
pydantic
python-multipart   
transformers>=4.36.0
torch
accelerate